Add files via upload
These files contain the RMarkdown code for the AERMOD output upload, parameter generation, risk assessment, and sensitivity analysis for the evaluation of the risk of respiratory infection with Legionella pneumophila for individuals living downwind of the active irrigator spraying with diluted municipal wastewater. Also included are the RData files necessary for running the code, since the original data files are too big to upload.